added more test cases

+++ main.py
@@ -2,9 +2,17 @@ from chunk import parse_code, chunk_code
 from embedding import clear_collection
 import requests
 
-github_url = "https://github.com/arjunMandair17/RAG-Codebase-Query"
+github_url = "https://github.com/arjunMandair17/Job-Application-Tracker"
 
-clear_collection()
+
+
+confirm = input("Clear collection? (y/n): ")
+if confirm == "y":
+    clear_collection()
+
+confirm = input("Ingest codebase? (y/n): ")
+if confirm == "n":
+    exit()
 
 response = requests.delete("http://localhost:8000/ingest")
 print(response.json())
@@ -16,4 +24,3 @@ while True:
     query = input("Enter a query: ")
     response = requests.post("http://localhost:8000/retrieve", json={"query": query})
     print(response.json()["answer"])
-clear_collection()
